Add support for RouterOS 6.x splitted counters in interface
The newer RouterOS doesn't use the counters in format:
  Param 9: bytes = 1090924066,2872515632
but it is now splitted, like:
  Param 13: rx-byte = 9003084616
  Param 14: tx-byte = 37225808829

The change affected only the interface list (/interface/print), but not the
counters in registration table (/interface/wireless/registration-table/print).

This commit is trying to detect this and convert it accordingly.

Problem is also reported here:
http://forum.mikrotik.com/viewtopic.php?t=65227